Continue implementing index pattern.
