+++ src/components/main/index.js
@@ -1,4 +1,8 @@
 
 import MainContainer from 'src/components/main/ApolloMainContainer';
+import DashboardLinkOrButton from 'src/components/main/DashboardLinkOrButton';
+import LoginModal from 'src/components/main/LoginModal';
+
+export { DashboardLinkOrButton, LoginModal };
 
 export default MainContainer;
