+++ src/components/modules/dashboard/index.js
@@ -1,3 +1,6 @@
 import Dashboard from 'src/components/modules/dashboard/ApolloDashboard';
+import UserStatusesPresentation from 'src/components/modules/dashboard/UserStatusesPresentation';
+
+export { UserStatusesPresentation };
 
 export default Dashboard;
